Change how we define number of points
Defining the number of points for initialConds and physics loss is much better, because it doesn't scale crazily like the previous solution (of defining the discretization of t and kc), so we can train with higher discretizations of k and c even on larger problems. Results still look good. Also fixed some dangling floats used before that would fail some tests :)